RequestValTest sprinkled out across multiple directive tests
Client ip extraction test with incorrect header name fixed.
